requirements.txt
streamlit
Pillow
transformers
torch
sentencepiece
json